# openapi-gpt.yaml
openapi: 3.1.0
info:
  title: CRM Lead API
  version: 1.0.0
  description: Read-only live CRM lead lookup API.
servers:
  - url: https://crm-gpt-api.onrender.com
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Lead:
      type: object
      additionalProperties: true
      properties:
        id: {type: string, format: uuid}
        lead_number: {type: string}
        name: {type: [string, "null"]}
        phone: {type: [string, "null"]}
        email: {type: [string, "null"]}
        created_at: {type: string, format: date-time}
        lead_status: {type: [string, "null"]}
        stage: {type: [string, "null"]}
        stage_label: {type: [string, "null"]}
        source: {type: [string, "null"]}
        source_label: {type: [string, "null"]}
        program: {type: [string, "null"]}
        heat: {type: [string, "null"]}
        score: {type: [integer, "null"]}
        assigned_to: {type: [string, "null"]}
      required: [id, lead_number, created_at]
    LeadList:
      type: object
      properties:
        leads: {type: array, items: {$ref: "#/components/schemas/Lead"}}
        limit: {type: integer}
      required: [leads, limit]
security:
  - bearerAuth: []
paths:
  /health:
    get:
      operationId: healthCheck
      description: Check whether the CRM bridge is running.
      security: []
      responses:
        "200": {description: Healthy}
  /leads/latest:
    get:
      operationId: getLatestLeads
      description: Use for latest or most recently created leads.
      parameters:
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 10}}
        - {name: offset, in: query, description: Start position; follow next_offset from the previous response., schema: {type: integer, minimum: 0, default: 0}}
      responses:
        "200":
          description: Latest leads
          content: {application/json: {schema: {$ref: "#/components/schemas/LeadList"}}}
  /leads/count:
    get:
      operationId: getTotalLeadCount
      description: Get the total lead count before retrieving all paginated batches.
      responses:
        "200": {description: Total lead count}
  /leads/today:
    get:
      operationId: getTodaysLeads
      description: Use when the user asks to show leads created today.
      parameters:
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
      responses:
        "200": {description: Today's leads}
  /leads/count/today:
    get:
      operationId: getTodaysLeadCount
      description: Use when the user asks how many leads were created today.
      responses:
        "200": {description: Today's lead count}
  /leads/search:
    get:
      operationId: searchLeads
      description: Search live leads by name, phone, email, or lead number.
      parameters:
        - {name: query, in: query, required: true, schema: {type: string, minLength: 2, maxLength: 200}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
      responses:
        "200": {description: Matching leads}
  /leads/status/{status}:
    get:
      operationId: getLeadsByStatus
      description: Use for leads in a named status or stage, such as converted.
      parameters:
        - {name: status, in: path, required: true, schema: {type: string}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
      responses:
        "200": {description: Leads matching the status}
  /leads/assigned/{person}:
    get:
      operationId: getLeadsByAssignee
      description: Use when the user asks for leads assigned to a named person.
      parameters:
        - {name: person, in: path, required: true, schema: {type: string}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
      responses:
        "200": {description: Leads assigned to the matching person}
  /reports/today:
    get:
      operationId: getTodaysCRMReport
      description: Use first for today's update, daily report, or overall CRM summary.
      responses:
        "200": {description: Today's CRM counts}
  /learners/latest:
    get:
      operationId: getLatestLearners
      description: Get the most recently created learner enrolments.
      parameters:
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
      responses: {"200": {description: Latest learners}}
  /learners/today:
    get:
      operationId: getTodaysLearners
      description: Get learner enrolments created today.
      parameters:
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
      responses: {"200": {description: Today's learners}}
  /learners/search:
    get:
      operationId: searchLearners
      description: Search learners by name, phone, email, or enrolment number.
      parameters:
        - {name: query, in: query, required: true, schema: {type: string, minLength: 2, maxLength: 200}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
      responses: {"200": {description: Matching learners}}
  /messages/latest:
    get:
      operationId: getLatestMessages
      description: Get recent WhatsApp, email, or voice messages, optionally inbound or outbound.
      parameters:
        - {name: channel, in: query, schema: {type: string, enum: [whatsapp, email, voice]}}
        - {name: direction, in: query, schema: {type: string, enum: [inbound, outbound]}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
      responses: {"200": {description: Latest messages}}
  /messages/today:
    get:
      operationId: getTodaysMessages
      description: Get today's WhatsApp, email, or voice messages, optionally inbound or outbound.
      parameters:
        - {name: channel, in: query, schema: {type: string, enum: [whatsapp, email, voice]}}
        - {name: direction, in: query, schema: {type: string, enum: [inbound, outbound]}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
      responses: {"200": {description: Today's messages}}
  /conversations/latest:
    get:
      operationId: getLatestConversations
      description: Get recent conversations with contact, last-message, unread, and assignee context.
      parameters:
        - {name: channel, in: query, schema: {type: string, enum: [whatsapp, email, voice]}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
      responses: {"200": {description: Latest conversations}}
  /conversations/unread:
    get:
      operationId: getUnreadConversations
      description: Use whenever the user asks for unread conversations. Returns only conversations whose unread_count is greater than zero.
      parameters:
        - {name: channel, in: query, schema: {type: string, enum: [whatsapp, email, voice]}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 100}}
      responses: {"200": {description: Unread conversations ordered by unread count and recency}}
  /activities/today:
    get:
      operationId: getTodaysActivities
      description: Get today's operational CRM activity timeline.
      parameters:
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 50}}
      responses: {"200": {description: Today's CRM activities}}